Update README.md
In order to run this program, a Java editor like JCreator or the one I used, Dr.Java, has to be downloaded. Then copy the program on to a blank program space. Also, the DrawingPanel class must also be present when trying to run this program. The DrawingPanel class will also be in my account along with MathGame class.To run the program, either click on compile in Dr.Java at the top or click on build in JCreator. Then it schold be able to say "process completed", then click on run to start the program.